Disable db-reset in production

+++ webed/views/page.py
@@ -61,9 +61,12 @@ def main (page='home', template='index.html'):
         print >> sys.stderr, "Time Stamp: %s" % session['timestamp']
 
     if 'reset' in request.args: db_reset (); init ()
-    if 'refresh' in request.args: db_clean (); init ()
+    if 'refresh' in request.args: db_refresh (); init ()
 
     return render_template (template, page=page, debug=app.debug)
+
+###############################################################################
+###############################################################################
 
 @page.route ('/reset/')
 def reset ():
@@ -72,10 +75,16 @@ def reset ():
 
 @page.route ('/refresh/')
 def refresh ():
-    db_clean (); init ();
+    db_refresh (); init ();
     return jsonify (dict (success=True))
 
+###############################################################################
+###############################################################################
+
 def db_reset ():
+
+    if not app.debug or app.testing:
+        return
 
     db.drop_all ()
     db.create_all ()
@@ -84,7 +93,7 @@ def db_reset ():
     db.session.add (user)
     db.session.commit ()
 
-def db_clean ():
+def db_refresh ():
 
     if 'root_uuid' in session:
         base = Q (Node.query).one_or_default (uuid=session['root_uuid'])
